requirements.txt
customtkinter==5.2.2
keyboard==0.13.5
pystray==0.19.5
Pillow>=11.0.0
mouse==0.7.1
psutil>=5.9.0
